refactor: [#186427501] generalize dynamics API and update config variable names
