+++ .circleci/config.yml
@@ -205,10 +205,10 @@ commands:
             echo 'export USE_BASIC_AUTH=$USE_BASIC_AUTH_DEV' >> $BASH_ENV
             echo 'export BASIC_AUTH_USERNAME=$BASIC_AUTH_USERNAME' >> $BASH_ENV
             echo 'export BASIC_AUTH_PASSWORD=$BASIC_AUTH_PASSWORD' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_ORG_URL=$DCA_DYNAMICS_ORG_URL_DEV' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_CLIENT_ID=$DCA_DYNAMICS_CLIENT_ID_DEV' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_SECRET=$DCA_DYNAMICS_SECRET_DEV' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_TENANT_ID=$DCA_DYNAMICS_TENANT_ID_DEV' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_URL=$DYNAMICS_LICENSE_STATUS_URL_DEV' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_CLIENT_ID=$DYNAMICS_LICENSE_STATUS_CLIENT_ID_DEV' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_SECRET=$DYNAMICS_LICENSE_STATUS_SECRET_DEV' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_TENANT_ID=$DYNAMICS_LICENSE_STATUS_TENANT_ID_DEV' >> $BASH_ENV
             echo 'export FEATURE_DYNAMICS_PUBLIC_MOVERS=$FEATURE_DYNAMICS_PUBLIC_MOVERS_DEV' >> $BASH_ENV
 
   set-cypress-env-vars:
@@ -238,10 +238,10 @@ commands:
             echo 'export AWS_CRYPTO_CONTEXT_PURPOSE=$AWS_CRYPTO_CONTEXT_PURPOSE' >> $BASH_ENV
             echo 'export AWS_CRYPTO_CONTEXT_ORIGIN=$AWS_CRYPTO_CONTEXT_ORIGIN' >> $BASH_ENV
             echo 'export DISABLE_GTM=true' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_ORG_URL=$DCA_DYNAMICS_ORG_URL_DEV' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_CLIENT_ID=$DCA_DYNAMICS_CLIENT_ID_DEV' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_SECRET=$DCA_DYNAMICS_SECRET_DEV' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_TENANT_ID=$DCA_DYNAMICS_TENANT_ID_DEV' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_URL=$DYNAMICS_LICENSE_STATUS_URL_DEV' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_CLIENT_ID=$DYNAMICS_LICENSE_STATUS_CLIENT_ID_DEV' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_SECRET=$DYNAMICS_LICENSE_STATUS_SECRET_DEV' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_TENANT_ID=$DYNAMICS_LICENSE_STATUS_TENANT_ID_DEV' >> $BASH_ENV
             echo 'export FEATURE_DYNAMICS_PUBLIC_MOVERS=$FEATURE_DYNAMICS_PUBLIC_MOVERS_DEV' >> $BASH_ENV
 
   set-staging-env-vars:
@@ -282,10 +282,10 @@ commands:
             echo 'export USE_BASIC_AUTH=$USE_BASIC_AUTH_STAGING' >> $BASH_ENV
             echo 'export BASIC_AUTH_USERNAME=$BASIC_AUTH_USERNAME' >> $BASH_ENV
             echo 'export BASIC_AUTH_PASSWORD=$BASIC_AUTH_PASSWORD' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_ORG_URL=$DCA_DYNAMICS_ORG_URL_STAGING' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_CLIENT_ID=$DCA_DYNAMICS_CLIENT_ID_STAGING' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_SECRET=$DCA_DYNAMICS_SECRET_STAGING' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_TENANT_ID=$DCA_DYNAMICS_TENANT_ID_STAGING' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_URL=$DYNAMICS_LICENSE_STATUS_URL_STAGING' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_CLIENT_ID=$DYNAMICS_LICENSE_STATUS_CLIENT_ID_STAGING' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_SECRET=$DYNAMICS_LICENSE_STATUS_SECRET_STAGING' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_TENANT_ID=$DYNAMICS_LICENSE_STATUS_TENANT_ID_STAGING' >> $BASH_ENV
             echo 'export FEATURE_DYNAMICS_PUBLIC_MOVERS=$FEATURE_DYNAMICS_PUBLIC_MOVERS_STAGING' >> $BASH_ENV
 
   set-prod-env-vars:
@@ -324,10 +324,10 @@ commands:
             echo 'export AWS_CRYPTO_CONTEXT_PURPOSE=$AWS_CRYPTO_CONTEXT_PURPOSE' >> $BASH_ENV
             echo 'export AWS_CRYPTO_CONTEXT_ORIGIN=$AWS_CRYPTO_CONTEXT_ORIGIN' >> $BASH_ENV
             echo 'export OUTAGE_ALERT_CONFIG_URL=$OUTAGE_ALERT_CONFIG_URL' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_ORG_URL=$DCA_DYNAMICS_ORG_URL_PROD' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_CLIENT_ID=$DCA_DYNAMICS_CLIENT_ID_PROD' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_SECRET=$DCA_DYNAMICS_SECRET_PROD' >> $BASH_ENV
-            echo 'export DCA_DYNAMICS_TENANT_ID=$DCA_DYNAMICS_TENANT_ID_PROD' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_URL=$DYNAMICS_LICENSE_STATUS_URL_PROD' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_CLIENT_ID=$DYNAMICS_LICENSE_STATUS_CLIENT_ID_PROD' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_SECRET=$DYNAMICS_LICENSE_STATUS_SECRET_PROD' >> $BASH_ENV
+            echo 'export DYNAMICS_LICENSE_STATUS_TENANT_ID=$DYNAMICS_LICENSE_STATUS_TENANT_ID_PROD' >> $BASH_ENV
             echo 'export FEATURE_DYNAMICS_PUBLIC_MOVERS=$FEATURE_DYNAMICS_PUBLIC_MOVERS_PROD' >> $BASH_ENV
 
   set-content-env-vars:
