+++ api/src/client/dynamics/license-status/DynamicsLicenseApplicationIdClient.ts
